Avoid mixing signed/unsigned comparison.

+++ src/theft_call.c
@@ -57,7 +57,7 @@ theft_call(struct theft_run_info *run_info, void **args) {
 static void
 get_arg_hash_buffer(theft_hash *buffer,
         struct theft_run_info *run_info, void **args) {
-    for (int i = 0; i < run_info->arity; i++) {
+    for (uint8_t i = 0; i < run_info->arity; i++) {
         struct theft_type_info *ti = run_info->type_info[i];
         theft_hash h = ti->hash(args[i], ti->env);
         LOG(4, "%s: arg %d hash; 0x%016" PRIx64 "\n", __func__, i, h);
